test(coordinator): split stop-phase coverage
